Two threads with right comments and using methods wait and sleep.
